refactor: #22 커뮤니티 페이지 전체 퍼블리싱
