+++ src/app/community/[postId]/components/CommentForm.tsx
@@ -35,24 +35,15 @@ export const CommentForm = ({ newComment, onCommentChange, onCommentSubmit }: Co
           </div>
 
           {/* 댓글 입력 영역 */}
-          <div className="relative flex-1">
+          <div className="flex-1">
             <input
               type="text"
               placeholder="댓글을 입력하세요..."
               value={newComment}
               onChange={(e) => onCommentChange(e.target.value)}
               maxLength={500}
-              className="w-full rounded-full border border-gray-200 bg-gray-50 px-3 py-2 pr-16 text-sm placeholder-gray-500 transition-all focus:border-violet-500 focus:bg-white focus:outline-none focus:ring-2 focus:ring-violet-500/20 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-violet-400 dark:focus:bg-gray-600 sm:px-4 sm:py-3 sm:pr-20"
+              className="w-full rounded-full border border-gray-200 bg-gray-50 px-3 py-2 text-sm placeholder-gray-500 transition-all focus:border-violet-500 focus:bg-white focus:outline-none focus:ring-2 focus:ring-violet-500/20 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-violet-400 dark:focus:bg-gray-600 sm:px-4 sm:py-3"
             />
-
-            {/* 글자 수 카운터 */}
-            <div className="absolute right-12 top-1/2 -translate-y-1/2 sm:right-16">
-              <span
-                className={`text-xs ${newComment.length > 450 ? 'text-red-500' : 'text-gray-400'}`}
-              >
-                {newComment.length}/500
-              </span>
-            </div>
           </div>
 
           {/* 전송 버튼 */}
@@ -70,49 +61,12 @@ export const CommentForm = ({ newComment, onCommentChange, onCommentSubmit }: Co
           </button>
         </div>
 
-        {/* 추가 옵션 및 미리보기 */}
-        {newComment.length > 0 && (
-          <div className="mt-3 space-y-2 pl-14">
-            {/* @멘션 미리보기 */}
-            {newComment.includes('@') && (
-              <div className="rounded-lg bg-gray-50 p-3 dark:bg-gray-700/50">
-                <div className="mb-1 text-xs font-medium text-gray-500 dark:text-gray-400">
-                  미리보기:
-                </div>
-                <div className="text-sm text-gray-800 dark:text-gray-200">
-                  {renderMentionPreview(newComment)}
-                </div>
-              </div>
-            )}
-
-            <div className="flex items-center justify-between">
-              <div className="flex items-center gap-4 text-xs text-gray-500 dark:text-gray-400">
-                <button
-                  type="button"
-                  className="flex items-center gap-1 transition-colors hover:text-violet-600"
-                >
-                  <span>@</span>
-                  <span>멘션</span>
-                </button>
-                <button
-                  type="button"
-                  className="flex items-center gap-1 transition-colors hover:text-violet-600"
-                >
-                  <span>#</span>
-                  <span>태그</span>
-                </button>
-              </div>
-
-              <button
-                type="button"
-                onClick={() => onCommentChange('')}
-                className="text-xs text-gray-400 transition-colors hover:text-red-500"
-              >
-                취소
-              </button>
-            </div>
-          </div>
-        )}
+        {/* 글자 수 카운터 - 우측 하단으로 이동 */}
+        <div className="mt-2 flex justify-end">
+          <span className={`text-xs ${newComment.length > 450 ? 'text-red-500' : 'text-gray-400'}`}>
+            {newComment.length}/500
+          </span>
+        </div>
       </form>
     </div>
   );
