+++ src/app/community/components/EmptyState/index.ts
@@ -1,0 +1,1 @@
+export { EmptyState } from './EmptyState';
