+++ src/app/community/components/FeaturedCard/index.ts
@@ -1,0 +1,1 @@
+export { FeaturedCard } from './FeaturedCard';
