+++ src/app/community/components/PostCard/index.ts
@@ -1,0 +1,1 @@
+export { PostCard } from './PostCard';
