+++ src/app/community/components/SortDropdown/SortDropdown.tsx
@@ -1,25 +1,20 @@
 'use client';
 import { useState, useRef, useEffect } from 'react';
-import { ChevronDown, Clock, Heart, Eye } from 'lucide-react';
+import { ChevronDown } from 'lucide-react';
 import { SortOption } from '../../types/community.types';
+import { SORT_OPTIONS } from '../../constants';
 
 interface SortDropdownProps {
   sortBy: SortOption;
   onSortChange: (sort: SortOption) => void;
 }
 
-const sortOptions = [
-  { value: 'latest' as SortOption, label: '최신순', icon: Clock },
-  { value: 'popular' as SortOption, label: '인기순', icon: Heart },
-  { value: 'views' as SortOption, label: '조회순', icon: Eye },
-];
-
 export const CommunitySortDropdown = ({ sortBy, onSortChange }: SortDropdownProps) => {
   const [isOpen, setIsOpen] = useState(false);
   const dropdownRef = useRef<HTMLDivElement>(null);
 
-  const currentOption = sortOptions.find((option) => option.value === sortBy);
-  const CurrentIcon = currentOption?.icon || Clock;
+  const currentOption = SORT_OPTIONS.find((option) => option.value === sortBy);
+  const CurrentIcon = currentOption?.icon;
 
   useEffect(() => {
     const handleClickOutside = (event: MouseEvent) => {
@@ -46,7 +41,7 @@ export const CommunitySortDropdown = ({ sortBy, onSortChange }: SortDropdownProp
         className="flex min-w-[120px] items-center justify-between gap-2 rounded-xl bg-gray-100 px-4 py-3 text-sm font-medium text-gray-800 shadow-sm transition-all hover:bg-gray-200 hover:shadow-md focus:outline-none focus:ring-2 focus:ring-violet-500 dark:bg-gray-700 dark:text-gray-100 dark:hover:bg-gray-600"
       >
         <div className="flex items-center gap-2">
-          <CurrentIcon className="h-4 w-4" />
+          {CurrentIcon && <CurrentIcon className="h-4 w-4" />}
           <span>{currentOption?.label}</span>
         </div>
         <ChevronDown
@@ -58,7 +53,7 @@ export const CommunitySortDropdown = ({ sortBy, onSortChange }: SortDropdownProp
 
       {isOpen && (
         <div className="absolute right-0 top-full z-50 mt-2 w-full min-w-[140px] overflow-hidden rounded-xl border border-gray-100 bg-white shadow-lg dark:border-gray-700 dark:bg-gray-800">
-          {sortOptions.map((option) => {
+          {SORT_OPTIONS.map((option) => {
             const OptionIcon = option.icon;
             const isSelected = option.value === sortBy;
