+++ src/app/community/hooks/useCommunity.ts
@@ -2,6 +2,8 @@
 import { useState, useMemo } from 'react';
 import { SortOption, CommunityFilters, BoardType, TagType } from '../types/community.types';
 import { SAMPLE_POSTS, FEATURED_POSTS } from '../temp/community.temp';
+import { POSTS_PER_PAGE, FEATURED_POSTS_LIMIT } from '../constants';
+import { filterPosts, sortPosts } from '../utils/postUtils';
 
 export const useCommunity = () => {
   const [filters, setFilters] = useState<CommunityFilters>({
@@ -11,59 +13,23 @@ export const useCommunity = () => {
   });
 
   const [currentPage, setCurrentPage] = useState(1);
-  const postsPerPage = 6;
 
   // 필터링된 게시물들
   const filteredPosts = useMemo(() => {
-    let filtered = [...SAMPLE_POSTS];
-
-    // 게시판 필터
-    if (filters.board !== 'all') {
-      filtered = filtered.filter((post) => post.board === filters.board);
-    }
-
-    // 글머리 필터
-    if (filters.tag) {
-      filtered = filtered.filter((post) => post.tag === filters.tag);
-    }
-
-    // 검색어 필터
-    if (filters.searchQuery) {
-      filtered = filtered.filter(
-        (post) =>
-          post.title.toLowerCase().includes(filters.searchQuery.toLowerCase()) ||
-          post.description?.toLowerCase().includes(filters.searchQuery.toLowerCase()) ||
-          post.author.toLowerCase().includes(filters.searchQuery.toLowerCase()),
-      );
-    }
-
-    // 정렬
-    switch (filters.sortBy) {
-      case 'popular':
-        filtered.sort((a, b) => (b.likeCount || b.likes || 0) - (a.likeCount || a.likes || 0));
-        break;
-      case 'views':
-        filtered.sort((a, b) => (b.views || 0) - (a.views || 0));
-        break;
-      case 'latest':
-      default:
-        filtered.sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime());
-        break;
-    }
-
-    return filtered;
+    const filtered = filterPosts(SAMPLE_POSTS, filters);
+    return sortPosts(filtered, filters.sortBy);
   }, [filters]);
 
   // 페이지네이션
-  const totalPages = Math.ceil(filteredPosts.length / postsPerPage);
+  const totalPages = Math.ceil(filteredPosts.length / POSTS_PER_PAGE);
   const currentPosts = filteredPosts.slice(
-    (currentPage - 1) * postsPerPage,
-    currentPage * postsPerPage,
+    (currentPage - 1) * POSTS_PER_PAGE,
+    currentPage * POSTS_PER_PAGE,
   );
 
   // 추천 게시물 (좋아요 순 상위 8개)
   const featuredPosts = useMemo(() => {
-    return FEATURED_POSTS.slice(0, 8);
+    return FEATURED_POSTS.slice(0, FEATURED_POSTS_LIMIT);
   }, []);
 
   const setSearchQuery = (query: string) => {
