+++ src/app/community/index.ts
@@ -1,0 +1,6 @@
+// Community module main exports
+export * from './components';
+export * from './hooks';
+export * from './types/community.types';
+export * from './constants';
+export * from './utils';
